// packages/edge-login-ui-rn/src/types/Scenes.ts
export type SceneNames =
  | 'ChangePasswordScene'
  | 'ChangePinScene'
  | 'ChangeRecoveryScene'
  | 'NewAccountWelcomeScene'
  | 'NewAccountUsernameScene'
  | 'NewAccountPasswordScene'
  | 'NewAccountPinScene'
  | 'NewAccountTosScene'
  | 'NewAccountWaitScene'
  | 'NewAccountReviewScene'
  | 'LandingScene'
  | 'LoadingScene'
  | 'OtpScene'
  | 'OtpRepairScene'
  | 'PasswordScene'
  | 'PinScene'
  | 'RecoveryLoginScene'
  | 'ResecurePasswordScene'
  | 'ResecurePinScene'
  | 'SecurityAlertScene'